demo/requirements.txt
# # --------- pytorch --------- #
https://download.pytorch.org/whl/cpu/torch-1.11.0%2Bcpu-cp39-cp39-linux_x86_64.whl
https://download.pytorch.org/whl/cpu/torchvision-0.12.0%2Bcpu-cp39-cp39-linux_x86_64.whl

lightning[extra]

gradio==3.39.0

tiktoken==0.4.0

# # --------- hydra --------- #
hydra-core==1.3.2
hydra-colorlog==1.2.0

